@@ -6,15 +6,15 @@
   "dependencies": {
     "@emotion/react": "11.11.0",
     "@emotion/styled": "11.11.0",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.13.0",
+    "@icons/material": "^0.3.0",
+    "@mui/icons-material": "^5.15.8",
+    "@mui/lab": "^5.0.0-alpha.171",
+    "@mui/material": "^5.15.8",
+    "@mui/styles": "^5.15.8",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.0.4",
     "@testing-library/user-event": "^12.1.7",
-    "axios": "^0.21.1",
+    "axios": "^1.13.2",
     "date-fns": "^2.16.1",
     "echarts": "5.2.2",
     "echarts-for-react": "3.0.2",
@@ -23,6 +23,7 @@
     "i18next": "^19.8.2",
     "i18next-browser-languagedetector": "^6.0.1",
     "izitoast-react": "^1.0.19",
+    "jwt-decode": "^4.0.0",
     "linkify-html": "^4.2.0",
     "linkify-it": "^5.0.0",
     "linkify-react": "^4.2.0",
@@ -30,32 +31,45 @@
     "material-ui-color": "^1.2.0",
     "mic-recorder-to-mp3": "^2.2.2",
     "moment": "^2.29.4",
-    "mui-audio-player": "^1.0.1",
     "or": "^0.2.0",
     "qrcode.react": "^1.0.0",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-color": "^2.19.3",
     "react-csv": "^2.2.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
     "react-flatpickr": "^3.10.13",
     "react-gauge-chart": "^0.4.1",
     "react-modal-image": "^2.5.0",
     "react-player": "^2.12.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
     "react-toastify": "^6.0.9",
-    "react-trello": "^2.2.11",
     "recharts": "^2.0.2",
     "socket.io-client": "^4.7.3",
-    "jwt-decode": "^4.0.0",
     "use-sound": "^4.0.1",
     "yup": "^0.32.8"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts --openssl-legacy-provider build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
+    "lint:fix": "eslint \"src/**/*.{js,jsx,ts,tsx}\" --fix",
+    "postinstall": "patch-package"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^8.57.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "patch-package": "^8.0.1",
+    "vite": "^5.4.11"
+  },
+  "overrides": {
+    "nth-check": "2.1.1",
+    "postcss": "8.4.31",
+    "webpack-dev-server": "4.15.2"
   },
   "eslintConfig": {
     "extends": "react-app"
